first public commit!
First public commit of Vevacious, a C++ code that generates Python code
and interacts with HOM4PS2, PyMinuit, and CosmoTransitions to attempt
to find all the minima of a 1-loop effective potential energy function.
This is version 0.2.7 since there are some small planned features to
add and more testing to do before 1.0.0 will be released.